Fixed after database changes

+++ app/src/main/java/team2/grocerycartcalculator/CheckoutActivity.java
@@ -52,7 +52,7 @@ public class CheckoutActivity extends AppCompatActivity {
 
 
     private void populateFoodList(){
-        Map<Food, Integer> listmap =  groceryList.getFoodQuantities();
+        Map<Food, Double> listmap =  groceryList.getFoodQuantities();
         ArrayList<Food> foods = new ArrayList<>(listmap.keySet());
         for(int i=0; i<foods.size(); i++){
             foodList.add(new List_Item(foods.get(i), listmap.get(foods.get(i)), groceryListID));
